refactor: 기존 User 클래스의 패키지 위치를 변경
