+++ src/main/java/codesquad/springcafe/user/User.java
@@ -1,4 +1,4 @@
-package codesquad.springcafe.domain;
+package codesquad.springcafe.user;
 
 public class User {
 
@@ -18,31 +18,16 @@ public class User {
         return userId;
     }
 
-    public void setUserId(String userId) {
-        this.userId = userId;
-    }
-
     public String getEmail() {
         return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
     }
 
     public String getNickname() {
         return nickname;
     }
 
-    public void setNickname(String nickname) {
-        this.nickname = nickname;
-    }
-
     public String getPassword() {
         return password;
     }
 
-    public void setPassword(String password) {
-        this.password = password;
-    }
 }
